+++ include/ODMatrix.hpp
@@ -16,11 +16,11 @@
 
 class ODMatrix {
  private:
-    std::vector<std::vector<int>> matrix;
-    std::unordered_map<int, int> idToIndexMap;
+    std::vector<std::vector<float>> matrix;
+    std::unordered_map<int64_t, int> idToIndexMap;
 
  public:
     ODMatrix();
     void loadFromFile(const std::string& filename);
-    int getTravelTime(const int& id1, const int& id2);
+    int getTravelTime(const int64_t& id1, const int64_t& id2);
 };
